QueryUpdateService implementation for assay runs. Auditing to be added in next checkin.
Turn on edit links for microarray assay runs. Other assay types won't see the links.

Use delete() methods on ExpObject subclasses, instead of calling static methods on ExperimentService that take int rowId arguments.

Stop using "Run Properties" and "Batch Properties" nodes for assay metadata. Instead, add the columns directly to the run and batch tables. Keep a hidden version of the old columns around for backwards compatibility.

+++ experiment/src/org/labkey/experiment/api/ExpRunImpl.java
@@ -26,6 +26,7 @@ import org.labkey.api.security.permissions.DeletePermission;
 import org.labkey.api.settings.AppProps;
 import org.labkey.api.util.URLHelper;
 import org.labkey.api.view.ActionURL;
+import org.labkey.api.view.UnauthorizedException;
 import org.labkey.experiment.DotGraph;
 import org.labkey.experiment.ExperimentRunGraph;
 import org.labkey.experiment.controllers.exp.ExperimentController;
@@ -171,6 +172,10 @@ public class ExpRunImpl extends ExpIdentifiableEntityImpl<ExperimentRun> impleme
 
     public void delete(User user)
     {
+        if (getContainer().hasPermission(user, DeletePermission.class))
+        {
+            throw new UnauthorizedException();
+        }
         ExperimentServiceImpl.get().deleteExperimentRunsByRowIds(getContainer(), user, getRowId());
     }
